Enable external creation of Settings object
* Create a base MORB_SLAM::Settings class for reading from config files
* Create a MORB_SLAM::CameraSettings class so a camera settings object can be created externally and passed into MORB_SLAM::System
* Update existing testing/example programs using MORB_SLAM to follow this convention
