+++ Examples/Stereo-Inertial/run_websocket.sh
@@ -1,1 +1,1 @@
-gdb --args ../../bin/realsense_websocket_client ../../Vocabulary/ORBvoc.txt ./RealSense_D435i.yaml 
+gdb --args ../../bin/external_IMU_client ../../Vocabulary/ORBvoc.txt ./RealSense_D435i.yaml ./websocket.yaml
