+++ include/MORB_SLAM/System.h
@@ -35,7 +35,7 @@
 #include "MORB_SLAM/KeyFrameDatabase.h"
 #include "MORB_SLAM/ORBVocabulary.h"
 #include "MORB_SLAM/ImuTypes.h"
-#include "MORB_SLAM/Settings.h"
+#include "MORB_SLAM/CameraSettings.hpp"
 #include "MORB_SLAM/Camera.hpp"
 #include "MORB_SLAM/Packet.hpp"
 
@@ -49,7 +49,7 @@ class Atlas;
 class Tracking;
 class LocalMapping;
 class LoopClosing;
-class Settings;
+class CameraSettings;
 typedef std::shared_ptr<Tracking> Tracking_ptr;
 
 class System {
@@ -63,7 +63,7 @@ class System {
  public:
     
     // Initialize the SLAM system. It launches the Local Mapping, Loop Closing and Viewer threads.
-    System(const std::string &strVocFile, const std::string &strSettingsFile, const CameraType sensor);
+    System(const std::string &strVocFile, std::shared_ptr<CameraSettings> camSettings);
 
     // Proccess the given stereo frame. Images must be synchronized and rectified.
     // Input images: RGB (CV_8UC3) or grayscale (CV_8U). RGB is converted to grayscale.
@@ -99,7 +99,7 @@ class System {
     bool getHasMergedLocalMap();
     bool getIsDoneVIBA();
 
-    std::shared_ptr<Settings> getSettings() const;
+    std::shared_ptr<CameraSettings> getSettings() const;
 
     bool getIsLoopClosed();
     void setIsLoopClosed(bool isLoopClosed);
@@ -154,7 +154,7 @@ private:
 
     std::string mStrVocabularyFilePath;
 
-    std::shared_ptr<Settings> settings;
+    std::shared_ptr<CameraSettings> settings;
 
 };
 typedef std::shared_ptr<System> System_ptr;
