+++ src/CameraSettings.cpp
@@ -1,142 +1,15 @@
-/**
- * This file is part of ORB-SLAM3
- *
- * Copyright (C) 2017-2021 Carlos Campos, Richard Elvira, Juan J. Gómez
- * Rodríguez, José M.M. Montiel and Juan D. Tardós, University of Zaragoza.
- * Copyright (C) 2014-2016 Raúl Mur-Artal, José M.M. Montiel and Juan D. Tardós,
- * University of Zaragoza.
- *
- * ORB-SLAM3 is free software: you can redistribute it and/or modify it under
- * the terms of the GNU General Public License as published by the Free Software
- * Foundation, either version 3 of the License, or (at your option) any later
- * version.
- *
- * ORB-SLAM3 is distributed in the hope that it will be useful, but WITHOUT ANY
- * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR
- * A PARTICULAR PURPOSE. See the GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along with
- * ORB-SLAM3. If not, see <http://www.gnu.org/licenses/>.
- */
-
-#include "MORB_SLAM/Settings.h"
-
-#include "MORB_SLAM/ImprovedTypes.hpp"
-#include <iostream>
-#include <opencv2/opencv.hpp>
-#include <opencv2/core/eigen.hpp>
-#include <opencv2/core/persistence.hpp>
+#include "MORB_SLAM/CameraSettings.hpp"
 
 #include "MORB_SLAM/CameraModels/KannalaBrandt8.h"
 #include "MORB_SLAM/CameraModels/Pinhole.h"
-#include <stdexcept>
+
+#include <opencv2/core/eigen.hpp>
+#include <opencv2/core/persistence.hpp>
 
 namespace MORB_SLAM {
-template <>
-bool Settings::readParameter<bool>(cv::FileStorage& fSettings, const std::string& name, bool& found, const bool required) {
-  cv::FileNode node = fSettings[name];
-  if (node.empty()) {
-    if (required) {
-      std::cerr << name << " required parameter does not exist, aborting..." << std::endl;
-      throw std::invalid_argument(name + " required parameter does not exist, aborting...");
-    } else {
-      std::cerr << name << " optional parameter does not exist..." << std::endl;
-      found = false;
-      return false;
-    }
-  } else if(node.isString()) {
-    found = true;
-    std::string s = node.string();
-    if(s=="y"||s=="Y"||s=="yes"||s=="Yes"||s=="YES"||s=="true"||s=="True"||s=="TRUE"||s=="on"||s=="On"||s=="ON")
-      return true;
-    else if(s=="n"||s=="N"||s=="no"||s=="No"||s=="NO"||s=="false"||s=="False"||s=="FALSE"||s=="off"||s=="Off"||s=="OFF")
-      return false;
-  }
-  throw std::invalid_argument(name + " bool setting was not set to a valid string");
-}
 
-template <>
-float Settings::readParameter<float>(cv::FileStorage& fSettings, const std::string& name, bool& found, const bool required) {
-  cv::FileNode node = fSettings[name];
-  if (node.empty()) {
-    if (required) {
-      std::cerr << name << " required parameter does not exist, aborting..." << std::endl;
-      throw std::invalid_argument(name + " required parameter does not exist, aborting...");
-    } else {
-      std::cerr << name << " optional parameter does not exist..." << std::endl;
-      found = false;
-      return 0.0f;
-    }
-  } else if (!node.isReal()) {
-    std::cerr << name << " parameter must be a real number, aborting..." << std::endl;
-    throw std::invalid_argument(name + " parameter must be a real number, aborting...");
-  } else {
-    found = true;
-    return node.real();
-  }
-}
 
-template <>
-int Settings::readParameter<int>(cv::FileStorage& fSettings, const std::string& name, bool& found, const bool required) {
-  cv::FileNode node = fSettings[name];
-  if (node.empty()) {
-    if (required) {
-      std::cerr << name << " required parameter does not exist, aborting..." << std::endl;
-      throw std::invalid_argument(name + " required parameter does not exist, aborting...");
-    } else {
-      std::cerr << name << " optional parameter does not exist..." << std::endl;
-      found = false;
-      return 0;
-    }
-  } else if (!node.isInt()) {
-    std::cerr << name << " parameter must be an integer number, aborting..." << std::endl;
-    throw std::invalid_argument(name + " parameter must be an integer number, aborting...");
-  } else {
-    found = true;
-    return node.operator int();
-  }
-}
-
-template <>
-std::string Settings::readParameter<std::string>(cv::FileStorage& fSettings, const std::string& name, bool& found, const bool required) {
-  cv::FileNode node = fSettings[name];
-  if (node.empty()) {
-    if (required) {
-      std::cerr << name << " required parameter does not exist, aborting..." << std::endl;
-      throw std::invalid_argument(name + " required parameter does not exist, aborting...");
-    } else {
-      std::cerr << name << " optional parameter does not exist..." << std::endl;
-      found = false;
-      return std::string();
-    }
-  } else if (!node.isString()) {
-    std::cerr << name << " parameter must be a std::string, aborting..." << std::endl;
-    throw std::invalid_argument(name + " parameter must be an integer number, aborting...");
-  } else {
-    found = true;
-    return node.string();
-  }
-}
-
-template <>
-cv::Mat Settings::readParameter<cv::Mat>(cv::FileStorage& fSettings, const std::string& name, bool& found, const bool required) {
-  cv::FileNode node = fSettings[name];
-  if (node.empty()) {
-    if (required) {
-      std::cerr << name << " required parameter does not exist, aborting..." << std::endl;
-      throw std::invalid_argument(name + " required parameter does not exist, aborting...");
-    } else {
-      std::cerr << name << " optional parameter does not exist..." << std::endl;
-      found = false;
-      return cv::Mat();
-    }
-  } else {
-    found = true;
-    return node.mat();
-  }
-}
-
-Settings::Settings(const std::string& configFile, const CameraType& sensor)
+CameraSettings::CameraSettings(const std::string& configFile, const CameraType& sensor)
     : sensor_(sensor),
       bNeedToUndistort_(false),
       bNeedToRectify_(false),
@@ -145,15 +18,7 @@ Settings::Settings(const std::string& configFile, const CameraType& sensor)
       thFarPoints_(0.f),
       activeLoopClosing_(true) {
 
-  // Open settings file
-  cv::FileStorage fSettings(configFile, cv::FileStorage::READ);
-  if (!fSettings.isOpened()) {
-    std::cerr << "[ERROR]: could not open configuration file at: " << configFile << std::endl;
-    std::cerr << "Aborting..." << std::endl;
-    throw std::invalid_argument("[ERROR]: could not open configuration file at: " + configFile);
-  } else {
-    std::cout << "Loading settings from " << configFile << std::endl;
-  }
+  cv::FileStorage fSettings = loadFile(configFile);
 
   // Read first camera
   readCamera1(fSettings);
@@ -196,7 +61,7 @@ Settings::Settings(const std::string& configFile, const CameraType& sensor)
   std::cout << "----------------------------------" << std::endl;
 }
 
-void Settings::readCamera1(cv::FileStorage& fSettings) {
+void CameraSettings::readCamera1(cv::FileStorage& fSettings) {
   bool found;
 
   // Read camera model
@@ -282,7 +147,7 @@ void Settings::readCamera1(cv::FileStorage& fSettings) {
   }
 }
 
-void Settings::readCamera2(cv::FileStorage& fSettings) {
+void CameraSettings::readCamera2(cv::FileStorage& fSettings) {
   bool found;
   std::vector<float> vCalibration;
   if (cameraModelType_ == PinHole) {
@@ -350,7 +215,7 @@ void Settings::readCamera2(cv::FileStorage& fSettings) {
   thDepth_ = readParameter<float>(fSettings, "Stereo.ThDepth", found);
 }
 
-void Settings::readImageInfo(cv::FileStorage& fSettings) {
+void CameraSettings::readImageInfo(cv::FileStorage& fSettings) {
   bool found;
   // Read original and desired image dimensions
   int originalRows = readParameter<int>(fSettings, "Camera.height", found);
@@ -402,7 +267,7 @@ void Settings::readImageInfo(cv::FileStorage& fSettings) {
   fps_ = readParameter<int>(fSettings, "Camera.fps", found);
 }
 
-void Settings::readIMU(cv::FileStorage& fSettings) {
+void CameraSettings::readIMU(cv::FileStorage& fSettings) {
   bool found;
   noiseGyro_ = readParameter<float>(fSettings, "IMU.NoiseGyro", found);
   noiseAcc_ = readParameter<float>(fSettings, "IMU.NoiseAcc", found);
@@ -415,7 +280,7 @@ void Settings::readIMU(cv::FileStorage& fSettings) {
   Tbc_ = Converter::toSophus(cvTbc);
 }
 
-void Settings::readRGBD(cv::FileStorage& fSettings) {
+void CameraSettings::readRGBD(cv::FileStorage& fSettings) {
   bool found;
 
   depthMapFactor_ = readParameter<float>(fSettings, "RGBD.DepthMapFactor", found);
@@ -424,7 +289,7 @@ void Settings::readRGBD(cv::FileStorage& fSettings) {
   bf_ = b_ * calibration1_->getParameter(0);
 }
 
-void Settings::readORB(cv::FileStorage& fSettings) {
+void CameraSettings::readORB(cv::FileStorage& fSettings) {
   bool found;
 
   nFeatures_ = readParameter<int>(fSettings, "ORBextractor.nFeatures", found);
@@ -434,7 +299,7 @@ void Settings::readORB(cv::FileStorage& fSettings) {
   minThFAST_ = readParameter<int>(fSettings, "ORBextractor.minThFAST", found);
 }
 
-void Settings::readViewer(cv::FileStorage& fSettings) {
+void CameraSettings::readViewer(cv::FileStorage& fSettings) {
   bool found;
 
   keyFrameSize_ = readParameter<float>(fSettings, "Viewer.KeyFrameSize", found);
@@ -452,14 +317,14 @@ void Settings::readViewer(cv::FileStorage& fSettings) {
   if (!found) imageViewerScale_ = 1.0f;
 }
 
-void Settings::readLoadAndSave(cv::FileStorage& fSettings) {
+void CameraSettings::readLoadAndSave(cv::FileStorage& fSettings) {
   bool found;
 
   sLoadFrom_ = readParameter<std::string>(fSettings, "System.LoadAtlasFromFile", found, false);
   sSaveto_ = readParameter<std::string>(fSettings, "System.SaveAtlasToFile", found, false);
 }
 
-void Settings::readOtherParameters(cv::FileStorage& fSettings) {
+void CameraSettings::readOtherParameters(cv::FileStorage& fSettings) {
   bool found;
 
   thFarPoints_ = readParameter<float>(fSettings, "System.thFarPoints", found, false);
@@ -474,7 +339,7 @@ void Settings::readOtherParameters(cv::FileStorage& fSettings) {
 
 }
 
-void Settings::precomputeRectificationMaps() {
+void CameraSettings::precomputeRectificationMaps() {
   // Precompute rectification maps, new calibrations, ...
   cv::Mat K1 = calibration1_->toK();
   K1.convertTo(K1, CV_64F);
@@ -513,11 +378,11 @@ void Settings::precomputeRectificationMaps() {
   }
 }
 
-std::ostream& operator<<(std::ostream& output, const Settings& settings) {
+std::ostream& operator<<(std::ostream& output, const CameraSettings& settings) {
   output << "SLAM settings: " << std::endl;
 
   output << "\t-Camera 1 parameters (";
-  if (settings.cameraModelType_ == Settings::PinHole || settings.cameraModelType_ == Settings::Rectified) {
+  if (settings.cameraModelType_ == CameraSettings::PinHole || settings.cameraModelType_ == CameraSettings::Rectified) {
     output << "Pinhole";
   } else {
     output << "Kannala-Brandt";
@@ -538,7 +403,7 @@ std::ostream& operator<<(std::ostream& output, const Settings& settings) {
   std::cout << "bout to start camera 2 stuff" << std::endl;
   if (settings.sensor_ == MORB_SLAM::CameraType::STEREO || settings.sensor_ == MORB_SLAM::CameraType::IMU_STEREO) {
     output << "\t-Camera 2 parameters (";
-    if (settings.cameraModelType_ == Settings::PinHole || settings.cameraModelType_ == Settings::Rectified) {
+    if (settings.cameraModelType_ == CameraSettings::PinHole || settings.cameraModelType_ == CameraSettings::Rectified) {
       output << "Pinhole";
     } else {
       output << "Kannala-Brandt";
@@ -574,7 +439,7 @@ std::ostream& operator<<(std::ostream& output, const Settings& settings) {
     }
     output << " ]" << std::endl;
 
-    if ((settings.sensor_ == MORB_SLAM::CameraType::STEREO || settings.sensor_ == MORB_SLAM::CameraType::IMU_STEREO) && settings.cameraModelType_ == Settings::KannalaBrandt) {
+    if ((settings.sensor_ == MORB_SLAM::CameraType::STEREO || settings.sensor_ == MORB_SLAM::CameraType::IMU_STEREO) && settings.cameraModelType_ == CameraSettings::KannalaBrandt) {
       output << "\t-Camera 2 parameters after resize: [ ";
       for (size_t i = 0; i < settings.calibration2_->size(); i++) {
         output << " " << settings.calibration2_->getParameter(i);
@@ -589,7 +454,7 @@ std::ostream& operator<<(std::ostream& output, const Settings& settings) {
     output << "\t-Stereo baseline: " << settings.b_ << std::endl;
     output << "\t-Stereo depth threshold : " << settings.thDepth_ << std::endl;
 
-    if (settings.cameraModelType_ == Settings::KannalaBrandt) {
+    if (settings.cameraModelType_ == CameraSettings::KannalaBrandt) {
       auto vOverlapping1 = std::static_pointer_cast<KannalaBrandt8>(settings.calibration1_)->getLappingArea();
       auto vOverlapping2 = std::static_pointer_cast<KannalaBrandt8>(settings.calibration2_)->getLappingArea();
       output << "\t-Camera 1 overlapping area: [ " << vOverlapping1[0] << " , " << vOverlapping1[1] << " ]" << std::endl;
@@ -618,4 +483,5 @@ std::ostream& operator<<(std::ostream& output, const Settings& settings) {
 
   return output;
 }
+
 };  // namespace MORB_SLAM
